requirements.txt
torch
gensim
numpy
matplotlib
scikit-learn
spacy
pandas
sentence-transformers
patsy
scipy
wordcloud
statsmodels
notebook
umap-learn
pyLDAvis